+++ mobile/native/src/trade/position/api.rs
@@ -28,12 +28,12 @@ pub enum PositionState {
 #[frb]
 #[derive(Debug, Clone)]
 pub struct Position {
-    pub leverage: f64,
-    pub quantity: f64,
+    pub leverage: f32,
+    pub quantity: f32,
     pub contract_symbol: ContractSymbol,
     pub direction: Direction,
-    pub average_entry_price: f64,
-    pub liquidation_price: f64,
+    pub average_entry_price: f32,
+    pub liquidation_price: f32,
     pub position_state: PositionState,
     pub collateral: u64,
 }
